+++ localhost_router.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-$requestUri = \urldecode(\parse_url((string)$_SERVER['REQUEST_URI'], PHP_URL_PATH));
+$requestUri = \urldecode(\parse_url($_SERVER['REQUEST_URI'] ?? '', PHP_URL_PATH));
 
 if ($requestUri !== '/' && \file_exists(__DIR__ . "/public$requestUri")) {
     return false;